WIP: refactor
