+++ src/controllers/survey-response.controller.ts
@@ -2,8 +2,8 @@ import { v4 as uuid } from 'uuid';
 import { None, Option } from 'funfix';
 import { SurveyResponse } from '../entities/survey-response';
 import { SurveyResponseRepository, SurveyId, SurveyResponseId } from '../types/survey';
-import { SurveyAnswer } from '../types/survey-answer';
-import { SubmissionId } from '../types/submission.types';
+import { SurveyAnswer } from '../entities/survey-answer';
+import { SubmissionId } from '../types/submission';
 import { Logger } from '@nestjs/common';
 
 export class SurveyResponseController {
